Moved activities to FunctionFragments package
