+++ app/src/main/java/com/pae14_1/FunctionFragments/RemoteControlActivity.java
@@ -1,9 +1,11 @@
-package com.pae14_1;
+package com.pae14_1.FunctionFragments;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
+
+import com.pae14_1.R;
 
 public class RemoteControlActivity extends AppCompatActivity {
 
@@ -13,7 +15,7 @@ public class RemoteControlActivity extends AppCompatActivity {
         setContentView(R.layout.activity_remote_control);
         setTitle("Remote Control");
 
-        setSupportActionBar((Toolbar) findViewById(R.id.mainToolbar));
+        setSupportActionBar((Toolbar) findViewById(R.id.remoteToolbar));
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
     }
