+++ app/src/main/java/com/pae14_1/FunctionFragments/SelfBalanceActivity.java
@@ -1,21 +1,25 @@
-package com.pae14_1;
+package com.pae14_1.FunctionFragments;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 
-public class InfraRedActivity extends AppCompatActivity {
+import com.pae14_1.R;
+
+public class SelfBalanceActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_infra_red);
-        setTitle(this.getString(R.string.infrared_robot));
+        setContentView(R.layout.activity_self_balance);
+        setTitle("Self Balance");
 
-        setSupportActionBar((Toolbar) findViewById(R.id.mainToolbar));
+        setSupportActionBar((Toolbar) findViewById(R.id.selfToolbar));
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+
     }
+
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case android.R.id.home:
