+++ app/src/main/java/com/pae14_1/FunctionFragments/UltraSonicActivity.java
@@ -1,23 +1,25 @@
-package com.pae14_1;
+package com.pae14_1.FunctionFragments;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 
-public class SelfBalanceActivity extends AppCompatActivity {
+import com.pae14_1.R;
+
+public class UltraSonicActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_self_balance);
-        setTitle("Self Balance");
+        setContentView(R.layout.activity_ultra_sonic);
+        setTitle(this.getResources().getString(R.string.ultrasonic_robot));
 
-        setSupportActionBar((Toolbar) findViewById(R.id.mainToolbar));
+        setSupportActionBar((Toolbar) findViewById(R.id.ultraToolbar));
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
     }
 
+    @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case android.R.id.home:
